% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENRT.R
\name{est.ENRT.sp}
\alias{est.ENRT.sp}
\title{Model spillover effect of ENRT data using mixed effect model}
\usage{
est.ENRT.sp(data)
}
\arguments{
\item{data}{The standard ENRT data with categorical covariate X and outcome y, created from the function \code{df.ENRT}}
}
\description{
Model spillover effect of ENRT data using mixed effect model
}
\details{
The following mixed effect model is applied to estimate the spillover effect of ENRT data
\deqn{Y_{ik}  = \sum_{h = 1}^{H} \zeta_{h} S_{kh} + \sum_{h = 1}^{H} \delta_{h} G_{ik}S_{kh} + u_k + \epsilon_{ik}, i = 2,\dots, n+1, k= 1,\dots, K,}
where \eqn{Y_{ik}} is the outcome for node \eqn{ik} in network \eqn{k}, \eqn{S_{kh}} is the category indicator for index participant in network \eqn{k} belonging to category \eqn{h},
\eqn{G_{ik}} is the number of treated network members for node \eqn{ik},
\eqn{\delta_h} represents the spillover effect of a network member whose index participant belongs to category \eqn{h}.
\eqn{u_k} is the random effect for networks and \eqn{\epsilon_{ik}} is the random noise.
}
\examples{
# simulation
n_node <- 1000
K <- 100
nodeid <- 1:n_node
netid <- 1:K
# ENRT design
df <- data.frame(nodeid=nodeid, idx_net=idx_net) \%>\% arrange(idx_net, by_group=TRUE)
# add R
df <- df \%>\% group_by(idx_net) \%>\% mutate(R= ifelse(row_number()==1, 1, 0))
# add Z
df <- df \%>\% mutate(Z= ifelse(R==1, sample(c(0,1), 1, prob=c(0.6,0.4), replace=TRUE), 0)) # random treatment allocation
# add X
df$X <- sample(LETTERS[1:4], n_node, replace=TRUE, prob=c(0.1,0.2,0.3,0.4))
df.ENRT(df$nodeid, df$R, df$idx_net, df$Z, X=df$X)
# add y
data <- df.ENRT(df$nodeid, df$R, df$idx_net, df$Z, X=df$X)
S <- as.matrix(fastDummies::dummy_cols(df$X, remove_selected_columns=TRUE))
cat_neigh <- data$G * S
# parameter setup
zeta <- rep(1, 4)
delta <- rep(2, 4)
sigma2_u <- 36
sigma2_e <- 4
y <- as.vector(S \%*\% zeta) + as.vector(cat_neigh \%*\% delta) + rnorm(K, mean=0, sd=sqrt(sigma2_u)) +
 rnorm(n_node, mean=0, sd=sqrt(sigma2_e))
data_full <- df.ENRT(df$nodeid, df$R, df$idx_net, df$Z, X=df$X, y=y)
model <- est.ENRT.sp(data_full) # lmer model
summary(model)
vcov(model)[5:8,5:8]
# need to add Junhan's model for estimation of AIE and ASpE
}
\seealso{
\code{\link{df.ENRT}}
}
