% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junhan_paper.R
\name{K.HISp}
\alias{K.HISp}
\title{Sample size (number of networks) calculation for a series of hypothesis tests of the Average Individual Effect (AIE) and
the Average Spillover Effect (ASpE)}
\usage{
K.HISp(
  power,
  tau,
  delta,
  rho_y,
  p,
  n,
  var_y,
  alpha,
  test = c("HIE", "HSpE", "HISpJ", "HISpC", "HOE")
)
}
\arguments{
\item{power}{size of power}

\item{tau}{Average Individual Effect (AIE)}

\item{delta}{Average Spillover Effect (ASpE)}

\item{rho_y}{Intra-class Correlation (ICC)}

\item{p}{random treatment allocation probability}

\item{n}{network size (social network members), in the basic ENRT design we assume the network size is equal for all networks}

\item{var_y}{variance of the outcome}

\item{alpha}{significant level}

\item{test}{choose which tests should be presented, by default, all the five tests (HIE, HSpE, HISpJ, HISpC and HOE) are presented}
}
\value{
A vector with the following elements:
\itemize{
\item{K_HIE:} The minimal required number of networks for testing AIE
\item{K_HSpE:} The minimal required number of networks for testing ASpE
\item{K_HISpJ:} The minimal required number of networks for joint test of AIE and ASpE
\item{K_HISpC:} The minimal required number of networks for conjunctive test of AIE and ASpE
\item{K_HOE:} The minimal required number of networks for testing the overall effect
}
}
\description{
These hypothesis tests include the direct tests for AIE (HIE) and ASpE (HSpE) respectively, the conjunctive (HISpJ) and dis-conjunctive (HISpC) test, and the test of overall effect (HOE).
}
\details{
Let \eqn{\tau} and \eqn{\delta} be the AIE and ASpE respectively.

\itemize{
\item{HIE:} The HIE test is used for the purpose with \eqn{H_0: \tau=0} against the alternative hypothesis \eqn{H_a: \tau \neq 0}.
A two-sided Z-test statistic \eqn{T_{\tau} = \sqrt{K} (\hat{\tau}/\hat{\sigma}_{\tau})}, where \eqn{\sigma^2_{\tau}} is the asymptotic variance of \eqn{\hat{\tau}}
(see equation (5) in the reference paper).
\item{HSpE:} The HSpE test is used for the purpose with
\eqn{H_0: \delta=0} against the alternative hypothesis \eqn{H_a: \delta \neq 0}.
A two-sided Z-test statistic \eqn{T_{\delta} = \sqrt{K} (\hat{\delta}/\hat{\sigma}_{\delta})}, where \eqn{\sigma^2_{\delta}} is the asymptotic variance of \eqn{\hat{\delta}}
(see equation (6) in the reference paper).
\item{HISpJ:} The HISpJ test is used for the purpose with
\eqn{H_0: \tau=0\  \text{and}\  \delta=0} against the alternative hypothesis \eqn{H_a: \tau \neq 0 \ \text{or}\ \delta \neq 0}.
A Wald test statistic \eqn{Q_J = K\hat{\bm{\theta}}^T \Sigma_{\tau\delta}^{-1} \hat{\bm{\theta}}}
is used and it follows a non-central \eqn{\chi^2} distribution (see equation (7) in the reference paper).
\item{HISpC:} The HISpC test is used for the purpose with
\eqn{H_0: \tau=0\  \text{or}\  \delta=0} against the alternative hypothesis \eqn{H_a: \tau \neq 0 \ \text{and}\ \delta \neq 0}.
A bivariate test statistic \eqn{Q_C = (T_{\tau},T_{\delta})} is used and it follows a bivariate normal distribution (see equation (8) in the reference paper).
\item{HOE:} The HOE test is used for the purpose with
\eqn{H_0: (\tau + n\delta)/(n+1)=0} against the alternative hypothesis \eqn{H_a: (\tau + n\delta)/(n+1) \neq 0}.
A two-sided Z-test statistic \eqn{T_{O} = \sqrt{K} \hat{\sigma}_{O} (\hat{\tau} + n\hat{\delta})/(n+1) }, where \eqn{\sigma^2_{O}} is the asymptotic variance of \eqn{(\hat{\tau} + n\hat{\delta})/(n+1)}
(see equation (9) in the reference paper).
}
}
\examples{
alpha <- 0.05 # significant level
power <- 0.80 # power
tau <- 1  # effect size for testing the Average Individual Effect (AIE)
delta <- 1 # effect size for testing the Average Spillover Effect (ASpE)
n <- c(1, 2, 5,10) # network size
ICC <- seq(0.1, 0.9, by=0.1) # intra-class correlation
var_y <- 1 # variance of outcome
K.HISp(power=power, tau=tau, delta=delta, rho_y=0.1, p=0.5, n=5, var_y=var_y, alpha=alpha)
df <- expand.grid(ICC, n); colnames(df) <- c("ICC", "n")
out <- cbind(df, t(apply(df, 1, function(x) K.HISp(power=power, tau=tau, delta=delta, rho_y=x[1], p=0.5, n=x[2], var_y=var_y, alpha=alpha), simplify=TRUE)))
print(out)
}
\references{
Design of egocentric network-based studies to estimate causal effects under interference (\url{https://arxiv.org/pdf/2308.00791})
}
\seealso{
\code{\link{n.HISp}}
}
