% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcb_ci.R
\name{mcbTest}
\alias{mcbTest}
\title{Multiple comparison with the best (MCB) test}
\usage{
mcbTest(
  aov_obj,
  conf.level = 0.95,
  two.sided = FALSE,
  direction = "max",
  control.group = NULL,
  method = c("MCA", "MCC", "MCB"),
  ...
)
}
\arguments{
\item{aov_obj}{a model generated by \code{anova} or \code{lme4}, usually it is a regression model}

\item{conf.level}{confidence level}

\item{two.sided}{if true the two-sided distribution is considered, else the one-sided distribution is considered}

\item{direction}{by default the larger the better, if 'min', the smaller the better, only used for MCB}

\item{control.group}{a string indicating the name of the control group, only used for}

\item{method}{method of multiple comparison adjustment.}
}
\value{
A list of
\itemize{
\item{ci} simultaneous confidence intervals
\item{p-value} both the overall p-value for heterogeneity test (\eqn{\delta_1 = \delta_2 =\dots =\delta_H}) and the adjusted p-value for each individual comparison (\eqn{\delta_h - \max_{j\neq h} \delta_j})
\item{cirtical.value} critical value
\item{bestgroup} A set of the best groups
}
}
\description{
For testing the heterogeneity of estimates across all groups, MCB simultaneously tests the difference
between the vale of each group and the best of other groups, such as \eqn{\delta_h -\max_{j\neq h}\delta_j}.
By construction of simultaneous CIs for estimates, a statistical inference decision is made based on the upper bounds or lower bounds.
Meanwhile, this function can also produce the results of multiple comparison with all-pairs (MCA) and the multiple comparison with the control group (MCC).
}
\examples{
## Data from Hsu's testbook on page 107
data <- data.frame(y = c(45, 59, 48, 46, 38, 47,
                        21, 12, 14, 17, 13, 17,
                        37, 32, 15, 25, 39, 41,
                        16, 11, 20, 21, 14, 7),
                  color = rep(c("yellow", "white", "red", "blue"), each = 6),
                  stringsAsFactors = TRUE)
fit <- aov(y ~ color, data = data)
mcbTest(fit,conf.level = 0.99, two.sided = FALSE, method = "MCB")
mcbTest(fit,conf.level = 0.99, method = "MCA")
# see page-107 maximum level is 'yellow'
mcbTest(fit,conf.level = 0.99, control.group = "yellow",two.sided = TRUE, method = "MCC")
}
\references{
Hsu, J. (1996). Multiple comparisons: theory and methods. CRC Press.
}
