% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power.size.mcb}
\alias{power.size.mcb}
\title{MCB standard power calculation based on given sample size and number of groups}
\usage{
power.size.mcb(n, alpha = 0.05, k, w)
}
\arguments{
\item{n}{the size of each group (currently we assume it is the same across all groups)}

\item{alpha}{significant level}

\item{k}{number of groups}

\item{w}{effect size in alternative hypothesis}
}
\value{
mcb power
}
\description{
MCB standard power calculation based on given sample size and number of groups
}
\examples{
n.sam = 2:12
xx = sapply(n.sam, power.size.mcb, k=8, w = 5/3, alpha = 0.05)
plot(n.sam,xx,type="l",col="red", ylab="Power", xlab="sample size")
abline(h = 1-0.05, lty=2,col="grey")
abline(h=0.90,lty=2,col="blue")
# simulation example in Hsu's textbook on page-241
alpha = 0.10
n.sam = seq(2, 24, by=1)
ratio = seq(0.8,1.2,by=0.1)
dat.pwr = matrix(NA,nrow = length(n.sam), ncol = length(ratio))
for (j in 1:length(ratio)) dat.pwr[,j] = sapply(n.sam, power.size.mcb, k=5, w = ratio[j], alpha = alpha)
plot(lowess(n.sam,dat.pwr[,1],f = 0.1), type="l",col=1,
     xlab = "sample size", ylab = "Power")
abline(h = 1-alpha, lty = 2, col="grey")
for (j in 2:length(ratio)) {
  lines(lowess(n.sam,dat.pwr[,j],f = 0.1),col=j)
}
legend("bottomright",legend = paste0("ratio = ", ratio), col = 1:length(ratio),
       lty = rep(1,length(ratio)), bty="n")
}
\references{
Hsu, J. (1996). Multiple comparisons: theory and methods. CRC Press.
}
