% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_mcb.R
\name{sam_size.mcb}
\alias{sam_size.mcb}
\title{Find minimal required number of networks per category for MCB}
\usage{
sam_size.mcb(
  n,
  w,
  H,
  p,
  rho_y,
  maxK = 1000,
  g_h = NULL,
  sigma2_y = 1,
  alpha = 0.05,
  beta = 0.1,
  isplot = FALSE,
  message = TRUE
)
}
\arguments{
\item{n}{number of non-center nodes in each network}

\item{w}{effect size for alternative hypothesis}

\item{H}{number of categories}

\item{rho_y}{ICC correlation}

\item{alpha}{Type I error, significant level}

\item{beta}{Type II error}

\item{message}{if true (default) approximate message of lambda is produced}

\item{is.plot}{if true plot the power vs. sample size}
}
\value{
numerical value of required number of networks
}
\description{
Find minimal required number of networks per category for MCB
}
\examples{
# a simple example
n = 2
w = 0.5
H = 3
p = 0.3
rho_y = 0.2
sam_size.mcb(n, w, H, p, rho_y = rho_y, isplot=TRUE)
}
\references{
Hsu, J. (1996). Multiple comparisons: theory and methods. CRC Press.
}
